build: initial commit.
The inital commit - hopefully works with minor modification.I can't
stand conventional commits, but some people soy for them, so I'll use
them. This commit sets up some infrastructure without documenting
it in the README - there's process that needs to be defined and
explained, too.

We have a monorepo powered by uv - the root directory contains the
dev dependencies while project directories define their own direct
dependencies. The uv lockfile is at the project root to keep every
library and application in sync.

This monorepo will also contain Terraform code and Kubernetes
manifests, as well as tooling to help deploy those. For that, the
.github/CODEOWNERS file contains myself (ff-dh) and James Mcminn
(jamesmcminn) as owners of the infrastructure and deployment
directories so that if there's a change there, we're marked as
reviewers.

Everything is licensed under MIT. Respect to the following for helping
me put this together:
https://github.com/carderne/postmodern-mono/
https://github.com/JasperHG90/uv-monorepo/