feat(tubescraper): start implementing the build.
As it stands, to build an image you add it to the compose file and
docker/podman-compose build it. Tubescraper has a dependency on the
"base_image" being built.  The builds have to re-lock the uv lockfile
because they strip out the unused dependencies, like the dev dependencies
at the top level.  Currently requires hard linking the uv lockfile to
the application directory, manually.